Fix the types
`WSGIEnviron` and `StartResponse` are defined in
[typeshed](https://github.com/python/typeshed) and can't be imported
into code. This is currently causing a `ModuleNotFound` error. Because
we don't control the values passed to these arguments anyway, there
isn't much value in typing them.

The one thing we do control, though, is the return value. This was left
untyped. Adding a type here ensures that what we return is the correct
shape.