Restore app's argument types
f0a8eddefdad4e7874dae2b9ff6f212f3b941003 removed the types from `app`'s
arguments because the types themselves only exist within
[typeshed](https://github.com/python/typeshed). The only time they can
be imported it when type checking the code. Fortunately there's a
convention for doing this sort of thing. The types are being added back
with the import behind an if that will never be true at runtime.