Use Python 3.8.1
This project currently uses Python 3.6.10 because that's what Heroku
uses by default. This version now only receives security fixes. This
change updates us to the latest version, 3.8.1.

It also updates the version checked by black and includes a file to
specify the version for users of pyenv (previously this file had been
included in the `.gitignore` file).

Closes #6