Only support numeric subdomains
This project will redirect to whatever subdomain is specified in the
`PYGOTHAM_YEAR` environment variable. A bad value (e.g., `example.com/`)
could send visitors to an unexpected destination.

Rather than doing that, we should take advantage of the fact that
`pygotham.org` will redirect to a year-specific site and only support
numeric values, falling back to the current year whenever the value
isn't found or is non-numeric.

Closes #7